+++ changelog.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Compress `outDir` to `zip`, `tar` and `tgz` formats.
 - Specify log level via `options.logLevel`.
-- Add platform, arch, Node and NW.js info in debug log
+- Add platform, arch, Node and NW.js info in debug log.
 
 ## [4.2.8] - 2023-06-29